+++ api/src/wrapper/json_check.py
@@ -62,6 +62,16 @@ def json_iterator(json):
     return missing
 
 
+# Call Schema Json to build Form JSON
+def set_form_json(query):
+    table_manager = TableManager(TABLES_PATH)
+    selected_table, form_json, token_tracker = request_tables_to_lm_from_db(query, table_manager, {})
+    form_json = {'cube': selected_table.name}
+    
+    return form_json
+
+
+
 if __name__ == "__main__":
     form_json = {
     "base_url": "https://api-dev.datausa.io/tesseract/data.jsonrecords?",
